Pass labels from asset list to asset-list-item
labels can not be changed and the items always use the default values. They need to be passed from the list. This is the quick solution i found but there might be more elegant ones. 

besides, I'm thinking button intent should be customizable too... Or the entire button. I'm open to hear sugestions and if I have time I'll try to submit a PR.